fix: Fix failing NewPipelineVersion tests
THIS COMMIT BREAKS A BUNCH OF OTHER TESTS

This example removes `t` from `PageProps` and adds it back as a
dependency for each component that extends `Page`. It also shows two
approaches to fixing tests that throw due to `t` being undefined
during prop initialization:

* Initialize `toolbarProps` to a default value instead of trying to use
  a temporary instance of the component
* Make `TestUtils.generatePageProps` accept additional props (instead
  of just `PageProps`) so the temporary component can be initialized
  with all dependencies (including `t`).

I'm not sure why `generatePageProps` is pre-initializing the toolbar
props with the component in the first place.